#2 The forbidden is working for /users and /users/ paths
