+++ user-profile/src/main/kotlin/systemkern/profile/UserRepository.kt
@@ -1,53 +1,20 @@
 package systemkern.profile
 
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.context.properties.ConfigurationProperties
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
 import org.springframework.data.repository.CrudRepository
-import org.springframework.data.rest.core.annotation.HandleBeforeCreate
-import org.springframework.data.rest.core.annotation.RepositoryEventHandler
 import org.springframework.data.rest.core.annotation.RepositoryRestResource
-import org.springframework.data.rest.core.config.RepositoryRestConfiguration
-import org.springframework.data.rest.webmvc.config.RepositoryRestConfigurer
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder
-import org.springframework.stereotype.Component
+import org.springframework.security.access.annotation.Secured
 import java.util.*
+import org.springframework.web.bind.annotation.GetMapping
 
 
 @RepositoryRestResource
-internal interface UserRepository : CrudRepository<User, UUID>
-//First push, it is done just to create the #2-add-spring-security branch in remote repository.
+internal interface UserRepository : CrudRepository<User, UUID> {
 
+    @Override
+    fun findByNameAndPassword(name: String,
+                              password: String): List<User>
 
-@Component
-@RepositoryEventHandler(User::class)
-internal class UserEventHandler(
-    @Autowired
-    internal val passwordEncoder: BCryptPasswordEncoder,
-    internal val userRepository: UserRepository
-) {
-    @HandleBeforeCreate
-    fun handleUserCreate(user: User) {
-        user.password = passwordEncoder.encode(user.password)
-    }
-}
-
-
-@Configuration
-internal class RepositoryRestConfig : RepositoryRestConfigurer {
-    override fun configureRepositoryRestConfiguration(config: RepositoryRestConfiguration) {
-        config.exposeIdsFor(User::class.java)
-    }
-}
-
-
-@Configuration
-@ConfigurationProperties("user-profile")
-internal class BCryptPasswordEncoderConfiguration {
-    var bcryptEncodeRounds: Long = 10
-
-    @Bean fun createBeanFoo() =
-        BCryptPasswordEncoder(10)
-
+    @Secured("ROLE_USER")
+    @GetMapping("{id}")
+    override fun findById(Id: UUID): Optional<User>
 }
